Fix double-to-float truncation in EulerAnglesToRotationMatrix
The function took float parameters but was called with doubles, silently
truncating precision. Per-iteration orthogonality drift drops from ~1e-7
to ~1e-16 with this change.

+++ c++/src/corrpts.cpp
@@ -88,7 +88,7 @@ void CorrPts::Reject(const double &min_planarity)
   dists_ = dists_new;
 }
 
-Eigen::Matrix3d EulerAnglesToRotationMatrix(float alpha1, float alpha2, float alpha3)
+Eigen::Matrix3d EulerAnglesToRotationMatrix(double alpha1, double alpha2, double alpha3)
 {
   Eigen::Matrix3d R;
   R <<
